+++ gradle.properties
@@ -1,8 +1,8 @@
 grpcVersion=1.68.1
 protobufVersion=4.28.3
 
-exileapiProtobufVersion=34.1.0.1.20260408174208.f723bbfb92f3
-exileapiGrpcVersion=1.80.0.1.20260408174208.f723bbfb92f3
+exileapiProtobufVersion=34.1.0.1.20260408183143.85794b77f79d
+exileapiGrpcVersion=1.80.0.1.20260408183143.85794b77f79d
 
 org.gradle.jvmargs=-Xmx4G
